Ajout d'un dossier initialisation pour l'algo génétique dans la branch principale
