+++ methaeuristics/Code/init_genetic/readme.txt
@@ -1,0 +1,2 @@
+init_genetic contient la fonction qui initialise un nombre n_sol de solutions.
+Le reste des fichiers sont des dépendancies.
